[TASK] Refractor USB audio subsystem:
	* Remove useless buffer allocation
	* Delete stubs and useless global variable
	* Delete useless scr and inc source file
